+++ Functions/Public/Get-PSIISSite.ps1
@@ -11,6 +11,11 @@
     Specify the name of the Application Pool to search for.
 .PARAMETER State
     Specify the state of the application pool to query.
+.PARAMETER Credential
+    The credentials to use for the connection. If not specified the connection will use the current user.
+    You can provide a PSCredential object, or use `New-PSIISSession` to create a PSCredential object that lives for the current powershell session.
+
+    See `Get-Help New-PSIISSession` for more details.
 .EXAMPLE
     Get-PSIISSite -ComputerName "localhost" -Name "DefaultSite"
 .NOTES
@@ -30,7 +35,9 @@ Function Get-PSIISSite() {
 
         [Parameter()]
         [ValidateSet('Started','Stopped', '*')]
-        [System.String] $State = '*'
+        [System.String] $State = '*',
+
+        [PSCredential]$Credential = $script:PSIISCredential
     )
 
     Begin {
@@ -68,7 +75,13 @@ Function Get-PSIISSite() {
             If (IsLocal $_) {
                 & $ScriptBlock -Name $Name -State $State
             } Else {
-                Invoke-Command -ComputerName $_ -ScriptBlock $ScriptBlock -ArgumentList $Name, $State | Select-Object * -ExcludeProperty RunspaceID
+                $InvokeCommandSplat = @{
+                    ComputerName = $_
+                    ScriptBlock = $ScriptBlock
+                    ArgumentList = @($Name, $State)
+                }
+                If ($null -ne $Credential) { $InvokeCommandSplat['Credential'] = $Credential }
+                Invoke-Command @InvokeCommandSplat | Select-Object * -ExcludeProperty RunspaceID
             }
         }
     }