+++ Functions/Public/Restart-PSIISSite.ps1
@@ -11,6 +11,11 @@
     Specify the website name to restart.
 .PARAMETER PassThru
     If true, the command will return the IIS information.
+.PARAMETER Credential
+    The credentials to use for the connection. If not specified the connection will use the current user.
+    You can provide a PSCredential object, or use `New-PSIISSession` to create a PSCredential object that lives for the current powershell session.
+
+    See `Get-Help New-PSIISSession` for more details.
 .EXAMPLE
     PS> Restart-PSIISSite -ComputerName WebServer01 -Name DefaultSite
 
@@ -36,7 +41,9 @@ Function Restart-PSIISSite() {
         [Alias('Sitename')]
         [System.String] $Name,
 
-        [switch]$PassThru
+        [switch]$PassThru,
+
+        [PSCredential]$Credential = $script:PSIISCredential
     )
 
     Begin {}
@@ -88,7 +95,13 @@ Function Restart-PSIISSite() {
                 If (IsLocal $_) {
                     & $ScriptBlock -Site $Site -PassThru:$PassThru
                 } Else {
-                    Invoke-Command -ComputerName $_ -ScriptBlock $ScriptBlock -ArgumentList $Site, $PassThru | Select-Object -ExcludeProperty RunspaceId
+                    $InvokeCommandSplat = @{
+                        ComputerName = $_
+                        ScriptBlock = $ScriptBlock
+                        ArgumentList = @($Site, $PassThru)
+                    }
+                    If ($null -ne $Credential) { $InvokeCommandSplat['Credential'] = $Credential }
+                    Invoke-Command @InvokeCommandSplat | Select-Object * -ExcludeProperty RunspaceID
                 }
             }
         }