+++ CHANGELOG.md
@@ -2,14 +2,34 @@
 
 ## v0.0.3
 
+`New Functions`:
+
++ `New-PSIISSession`
+  + This function will create a new session credential that all PSIIS cmdlets will reference for the life of the session (or until the credential is removed, see `Remove-PSIISSession`).
++ `Remove-PSIISSession`
+  + This function will remove a session credential created by `New-PSIISSession`.
+
+`Current Function Changes`:
+
+Each PSIIS command has been updated to accept a `-Credential` `[PSCredential]` object (with the exception of `Remove-PSIISSession`).
+You are now able to manually provide your own credential to each cmdlet, or invoke `New-PSIISSession` to create a session credential that all cmdlets will reference for the life of the session.
+
 `Additional Test Coverage`:
 
 + Adding additional test coverage, ensuring compatibility with Pester v5+.
++ Additional tests to ensure all public files are exported in the module manifest.
 
 `BUG Fixes`:
 
-+ [Issue #2](https://github.com/matthewjdegarmo/PSIISHelper/issues/2)
++ [Issue #2](https://github.com/matthewjdegarmo/PSIISHelper/issues/2) - [@thefirstnoe1](https://github.com/thefirstnoe1)
   + Fixed function returning `True` when a `-ComputerName` contained a period (eg. FQDNs)
+
+`Closed Issues`:
+
++ [Issue #2](https://github.com/matthewjdegarmo/PSIISHelper/issues/2) - [@thefirstnoe1](https://github.com/thefirstnoe1)
+  + Fixed function returning `True` when a `-ComputerName` contained a period (eg. FQDNs)
++ [Issue #3](https://github.com/matthewjdegarmo/PSIISHelper/issues/3) - [@thefirstnoe1](https://github.com/thefirstnoe1)
+  + Cmdlets now accept credentials using `-Credential` or `New-PSIISSession`.
 
 ## v0.0.2